+++ c/sort.cc
@@ -1213,7 +1213,7 @@ RowIndex DataTable::sortby(const std::vector<size_t>& colindices,
     if (out_grps) {
       *out_grps = Groupby::single_group(col0->nrows);
     }
-    return RowIndex::from_slice(i, col0->nrows, 1);
+    return RowIndex::from_slice(i, static_cast<int64_t>(col0->nrows), 1);
   }
   SortContext sc(nrows, col0->rowindex(),
                  (out_grps != nullptr) || (nsortcols > 1));
@@ -1231,7 +1231,7 @@ static RowIndex sort_tiny(const Column* col, Groupby* out_grps) {
   if (out_grps) {
     *out_grps = Groupby::single_group(col->nrows);
   }
-  return RowIndex::from_slice(i, col->nrows, 1);
+  return RowIndex::from_slice(i, static_cast<int64_t>(col->nrows), 1);
 }
